Tests: Implement in-memory worker pattern via simple WaitGroup
